pyproject.toml
[tool.poetry]
name = "ICH intership"
version = "0.1.0"
description = ""
authors = ["ICH"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
mysqlclient = "^2.2.4"
django-environ = "^0.11.2"
gunicorn = "^22.0.0"
drf-yasg = "^1.21.7"
setuptools = "^73.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.1.0"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
skip-string-normalization = true
target-version = ['py312']
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | venv
  | \.env
  | _build
  | buck-out
  | build
  | dist
  | .*/migrations
  | \.gitignore
  | __pycache__
  | .tox
  | .*.tox.*
)/
'''
